+++ konsave/__init__.py
@@ -1,9 +1,9 @@
-"""Top-level Konsave package."""
+'''Top-level Konsync package.'''
 
 from pkg_resources import get_distribution, DistributionNotFound
 
 try:
-    __version__ = get_distribution(__name__).version
+	__version__ = get_distribution(__name__).version
 except DistributionNotFound:
-    # Package is not installed
-    pass
+	# Package is not installed
+	pass